// packages/vendyx-admin-ui/src/lib/collections/components/collection-details/use-form/form-schema.ts
import z from 'zod';

import { CollectionContentType } from '@/lib/api/types';
import { FormMessages } from '@/shared/forms/form-messages';

export const CollectionDetailsFormSchema = z.object({
  name: z.string().min(1, FormMessages.required),
  description: z.string().optional(),
  image: z.instanceof(File).optional(),
  enabled: z.boolean(),
  contentType: z
    .enum([CollectionContentType.Products, CollectionContentType.Collections])
    .default(CollectionContentType.Products)
});